fix: :arrow_down: fix(package): Downgrade @livekit/components-react version (#233)
Description:

This commit downgrades the version of @livekit/components-react from 2.2.0 to 2.0.0 in the package-lock.json file. This change is necessary to ensure compatibility with other dependencies in the project.

Details:

The @livekit/components-react package was previously updated to version 2.2.0. However, this version was found to be incompatible with other dependencies in the project. As a result, the version of @livekit/components-react has been downgraded to 2.0.0.

Changeset:

The changeset involves modifying the version of @livekit/components-react in the package-lock.json file.

Modified files:

1. package-lock.json: The version of @livekit/components-react was changed from 2.2.0 to 2.0.0.

This commit ensures that the project's dependencies remain compatible with each other. It is recommended to test the project thoroughly after this change to ensure that no functionality has been affected.

Signed-off-by: RicardoGEsteves(Ricardo Esteves) 30535937+RicardoGEsteves@users.noreply.github.com

Refs: [#232]
Related to: [#232]